requirements.txt
feedparser
requests
beautifulsoup4
python-dotenv
google-genai
